Reverted change in f9259705b3e11fcd8d20e7a15a78f6f722563ad6, as it wasn't needed

+++ core/unit_test/TestPolicyConstruction.hpp
@@ -1181,8 +1181,8 @@ TEST(TEST_CATEGORY, md_range_policy_compile_time_deduction_guides) {
 
   constexpr int t[5]           = {};
   constexpr int64_t tt[5]      = {};
-  Kokkos::Array<int64_t, 3> a  = {};
-  Kokkos::Array<int64_t, 2> aa = {};
+  Kokkos::Array<int, 3> a  = {};
+  Kokkos::Array<int, 2> aa = {};
 
   Kokkos::MDRangePolicy<Kokkos::Rank<2>> ptc;
   Kokkos::MDRangePolicy pdc(ptc);
